+++ src/axmysql/messagestore/pkg.go
@@ -1,0 +1,3 @@
+// Package messagestore provides MySQL-specific implementations of the
+// interfaces in Ax's top-level "messagestore" package.
+package messagestore
