+++ src/axmysql/persistence/pkg.go
@@ -1,0 +1,3 @@
+// Package persistence provides MySQL-specific implementations of the interfaces
+// in Ax's top-level "persistence" package.
+package persistence
