+++ src/axmysql/sagasnapshot.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/jmalloc/ax/src/ax/persistence"
 	"github.com/jmalloc/ax/src/ax/saga"
+	mysqlpersistence "github.com/jmalloc/ax/src/axmysql/persistence"
 )
 
 // SnapshotRepository is an interface for loading and saving snapshots of
@@ -24,7 +25,7 @@ func (SnapshotRepository) LoadSagaSnapshot(
 		data        []byte
 	)
 
-	err := sqlTx(ptx).QueryRowContext(
+	err := mysqlpersistence.ExtractTx(ptx).QueryRowContext(
 		ctx,
 		`SELECT
 			instance_id,
@@ -67,7 +68,7 @@ func (SnapshotRepository) SaveSagaSnapshot(
 		return err
 	}
 
-	_, err = sqlTx(tx).ExecContext(
+	_, err = mysqlpersistence.ExtractTx(tx).ExecContext(
 		ctx,
 		`INSERT INTO saga_snapshot SET
 			instance_id = ?,
